Added surplus to gdp
Problem to polish in mid term future: surplus TS is extremely seasonal
for some reasons, decompose() do not cleanse away all the seasonal
component properly so that a part is still in the series. Next step
could consist in estimating an ARIMA and use it as a basis to clean out
seasonal component.